+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - App now opens instantly even without network connection
+- Better api connection status indicator in AppBar (orange if incompatible, MycelProject icon if reachable and compatible, spin animation)
 
 ## v0.0.2
 - Display informations on Mycel/Mycelium compatibility